fix(tools): update get_data_products_tool api call (#67)
This patch updates the `get_data_products` tool to use the current
format of the `/integration/data-products/v1/search-internally/{marketplace_id}/`
API endpoint. As of writing the endpoint now returns a payload in the
format of:
```json
{
    "search_id": "<uuid>",
    "results": []
}
```
whereas before it was simply the list of results.

This also changes the message to use a singular noun instead of plural
if there is only one data product.